requirements.txt
cryptography
pillow
numpy
opencv-python
pydub

